requirements.txt
pypdf==4.2.0
reportlab==4.1.0
PyYAML==6.0.1
jsonschema==4.22.0
streamlit==1.36.0
openai==1.40.0
python-dotenv==0.21.0
httpx==0.25.2
docling
